[Feat] 멤버 닉네임 설정(변경),삭제,로그아웃,로그인,Redis 연동 구현
Authorization 객체를 받아 인증하며 닉네임을 변경,삭제,로그아웃등을 실행하게 구현
로그인 시 RefreshToken이 Redis에 존재하지 않을경우 Redis에 생성되게끔 구현(확인완료)
로그아웃 시 AccessToken을 Redis에 BlackList로 올리는것 확인 필요(로그아웃 시 Header에 Authorization을 해독하여 가져옴)

issue:S08P12B111-325,S08P12B111-341,S08P12B111-346,S08P12B111-351,S08P12B111-361,S08P12B111-358

+++ back/src/main/java/com/udteam/miristock/entity/MemberEntity.java
@@ -19,7 +19,7 @@ public class MemberEntity {
     @Id
     @Column(name="member_no")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long memberNo;
+    private int memberNo;
 
     @Column(name="member_email",nullable = false,length = 320)
     private String memberEmail;
